feat: move scroll to top feature to separate component
